#34 refactor : 클래스명, 메서드명 수정
