+++ src/main/java/com/idea5/four_cut_photos_map/domain/favorite/dto/response/FavoriteResponseDto.java
@@ -1,6 +1,6 @@
 package com.idea5.four_cut_photos_map.domain.favorite.dto.response;
 
-import com.idea5.four_cut_photos_map.domain.shop.dto.response.ShopFavoritesResponseDto;
+import com.idea5.four_cut_photos_map.domain.shop.dto.response.ResponseFavoriteShop;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,5 +12,5 @@ import lombok.Setter;
 @Builder
 public class FavoriteResponseDto {
         private Long id;
-        private ShopFavoritesResponseDto shop;
+        private ResponseFavoriteShop shop;
 }
