+++ src/main/java/com/idea5/four_cut_photos_map/domain/shop/dto/KakaoKeywordResponseDto.java
@@ -1,4 +1,4 @@
-package com.idea5.four_cut_photos_map.domain.shop.dto.response;
+package com.idea5.four_cut_photos_map.domain.shop.dto;
 
 import lombok.*;
 
@@ -7,11 +7,9 @@ import lombok.*;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class KaKaoSearchResponseDto {
+public class KakaoKeywordResponseDto {
         public String place_name; // 장소명
         public String road_address_name; // 도로명 주소
         public String x; // 경도
         public String y; // 위도
-
-
 }
