+++ src/main/java/com/idea5/four_cut_photos_map/domain/shop/dto/response/ResponseFavoriteShop.java
@@ -1,19 +1,15 @@
 package com.idea5.four_cut_photos_map.domain.shop.dto.response;
 
-import com.idea5.four_cut_photos_map.domain.favorite.entity.Favorite;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-
-import java.util.List;
-
 @Getter
 @Setter
 @Builder
 @AllArgsConstructor
-public class ShopFavoritesResponseDto {
+public class ResponseFavoriteShop {
     private Long id;
     private String brand;
     private String name;
